fix: Rename EspClass.h to Esp.h and include from Arduino.h (#61)
Matches the real ESP32 Arduino Core where the file is Esp.h (not
EspClass.h). Arduino.h now includes Esp.h as the single entry point,
so ESP.getEfuseMac() and other ESP methods are available without
explicit includes in consumer code.
